Formatting changes, World Object base
Changed cache method names to be more understandable and added comments
Moved "App" to its own file and renamed it to ClientApp
Increased angle amount ( experimental )
Outline thickness can now be manually defined

Moved certain assets around in the filestructure
