+++ Circus/cache.py
@@ -5,8 +5,13 @@ class Cache:
     def __init__( self, app ):
         self.app = app
         self.app.done_counter = 0
-        self.stacked_sprite_cache = {}
+
+        # This dictionary contains all the raw layers of each stacked sprite
+        self.stacked_sprite_layer_cache = {}
+        # This dictionary contains the data for our sprites
         self.entity_sprite_cache = {}
+        
+        # Used in calculations, NUM_ANGLES defines how many possible angles a precompiled stacked sprite can have
         self.viewing_angle = 360 // NUM_ANGLES
         self.outline_thickness = 5
         self.alpha_value = 70  #
@@ -14,19 +19,27 @@ class Cache:
         #self.get_stacked_sprite_cache()
         #self.get_entity_sprite_cache()
 
-    def get_entity_sprite_cache( self ):
+    # This method should be called when a scene loads
+    # It takes all our sprite data and saves it into the cache
+    def cache_entity_sprite_data( self ):
+        # Go through each sprite entity
         for sprite_name in ENTITY_SPRITE_ATTRS:
+            # Create entry in our cache
             self.entity_sprite_cache[ sprite_name ] = {
                 'images': None
             }
+
+            # Get our entities attributes
             attrs = ENTITY_SPRITE_ATTRS[ sprite_name ]
-            images = self.get_layer_array( attrs )
+
+            images = self.create_stack_layer_array( attrs )
+            
             self.entity_sprite_cache[ sprite_name ][ 'images' ] = images
 
-            mask = self.get_entity_mask( attrs, images )
+            mask = self.create_entity_mask( attrs, images )
             self.entity_sprite_cache[ sprite_name ][ 'mask' ] = mask
 
-    def get_entity_mask( self, attrs, images ):
+    def create_entity_mask( self, attrs, images ):
         path = attrs.get( 'mask_path', False )
         if not path:
             return pg.mask.from_surface( images[ 0 ] )
@@ -36,61 +49,62 @@ class Cache:
             mask_image = pg.transform.scale( mask_image, vec2( mask_image.get_size() ) * scale )
             return pg.mask.from_surface( mask_image )
 
-    def get_stacked_sprite_cache( self ):
+    def cache_stacked_sprite_data( self ):
         for obj_name in STACKED_SPRITE_ATTRS:
-            self.stacked_sprite_cache[ obj_name ] = {
+            self.stacked_sprite_layer_cache[ obj_name ] = {
                 'rotated_sprites': {},
                 'alpha_sprites': {},
                 'collision_masks': {}
             }
             attrs = STACKED_SPRITE_ATTRS[ obj_name ]
-            layer_array = self.get_layer_array( attrs )
-            self.run_prerender( obj_name, layer_array, attrs )
+            layer_array = self.create_stack_layer_array( attrs )
+            self.compile_stacked_sprite_angles( obj_name, layer_array, attrs )
             self.app.done_counter += 1
             yield 1
         yield 'done'
 
-    def run_prerender( self, obj_name, layer_array, attrs ):
-        outline = attrs.get( 'outline', True )
+    def compile_stacked_sprite_angles( self, obj_name, layer_array, attrs ):
+        outline = attrs.get( 'outline', 0 )
         transparency = attrs.get( 'transparency', False )
         mask_layer = attrs.get( 'mask_layer', attrs[ 'num_layers' ] // 2 )
 
         for angle in range( NUM_ANGLES ):
-            surf = pg.Surface( layer_array[ 0 ].get_size() )
+            layer_size = layer_array[ 0 ].get_size()
+            surf = pg.Surface( layer_size )
             surf = pg.transform.rotate( surf, angle * self.viewing_angle )
             sprite_surf = pg.Surface( [ surf.get_width(), surf.get_height()
                                       + attrs[ 'num_layers' ] * attrs[ 'scale' ]] )
             sprite_surf.fill( 'khaki' )
             sprite_surf.set_colorkey( 'khaki' )
 
-            for ind, layer in enumerate( layer_array ):
+            for layer_index, layer in enumerate( layer_array ):
                 layer = pg.transform.rotate( layer, angle * self.viewing_angle )
-                sprite_surf.blit( layer, ( 0, ind * attrs[ 'scale' ] ))
+                sprite_surf.blit( layer, ( 0, layer_index * attrs[ 'scale' ] ))
 
                 # get collision mask
-                if ind == mask_layer:
+                if layer_index == mask_layer:
                     surf = pg.transform.flip( sprite_surf, True, True )
                     mask = pg.mask.from_surface( surf )
-                    self.stacked_sprite_cache[ obj_name ][ 'collision_masks' ][ angle ] = mask
+                    self.stacked_sprite_layer_cache[ obj_name ][ 'collision_masks' ][ angle ] = mask
 
             # get outline
             if outline:
                 outline_coords = pg.mask.from_surface( sprite_surf ).outline()
-                pg.draw.polygon( sprite_surf, 'black', outline_coords, self.outline_thickness )
+                pg.draw.polygon( sprite_surf, 'black', outline_coords, outline )
 
             # get alpha sprites
             if transparency:  #
                 alpha_sprite = sprite_surf.copy()
                 alpha_sprite.set_alpha( self.alpha_value )
                 alpha_sprite = pg.transform.flip( alpha_sprite, True, True )
-                self.stacked_sprite_cache[ obj_name ][ 'alpha_sprites' ][ angle ] = alpha_sprite
+                self.stacked_sprite_layer_cache[ obj_name ][ 'alpha_sprites' ][ angle ] = alpha_sprite
 
             image = pg.transform.flip( sprite_surf, True, True )
-            self.stacked_sprite_cache[ obj_name ][ 'rotated_sprites' ][ angle ] = image
+            self.stacked_sprite_layer_cache[ obj_name ][ 'rotated_sprites' ][ angle ] = image
 
 
 
-    def get_layer_array( self, attrs ):
+    def create_stack_layer_array( self, attrs ):
         # load sprite sheet
         sprite_sheet = pg.image.load( attrs[ 'path' ] ).convert_alpha()
         # scaling
