+++ Circus/draw_manager.py
@@ -1,0 +1,3 @@
+class DrawManager:
+    def __init__( self ):
+        pass
