+++ Circus/settings.py
@@ -11,7 +11,7 @@ PLAYER_SPEED = 0.4
 PLAYER_ROT_SPEED = 0.0015
 
 BG_COLOR = ( 9, 20, 38 ) #'white'  # olivedrab
-NUM_ANGLES = 72  # multiple of 360 -> 24, 30, 36, 40, 45, 60, 72, 90, 120, 180
+NUM_ANGLES = 360  # multiple of 360 -> 24, 30, 36, 40, 45, 60, 72, 90, 120, 180
 
 pg.mixer.init()
 pg.mixer.music.load( "assets/bgm/bgm.ogg" )
@@ -300,6 +300,7 @@ STACKED_SPRITE_ATTRS = {
         'y_offset': -130,
         'transparency': True,
         'mask_layer': 3,
+        'outline': 10,
     },
     'tree5': {
         'path': 'assets/stacked_sprites/tree5.png',
@@ -308,6 +309,7 @@ STACKED_SPRITE_ATTRS = {
         'y_offset': -130,
         'transparency': True,
         'mask_layer': 3,
+        'outline': 10,
     },
     'tree6': {
         'path': 'assets/stacked_sprites/tree6.png',
@@ -316,6 +318,7 @@ STACKED_SPRITE_ATTRS = {
         'y_offset': -130,
         'transparency': True,
         'mask_layer': 3,
+        'outline': 10,
     },
     'tree7': {
         'path': 'assets/stacked_sprites/tree7.png',
@@ -324,6 +327,7 @@ STACKED_SPRITE_ATTRS = {
         'y_offset': -130,
         'transparency': True,
         'mask_layer': 3,
+        'outline': 10,
     },
     'flower1': {
         'path': 'assets/stacked_sprites/flower1.png',
