+++ Circus/tilemap.py
@@ -1,3 +1,7 @@
+from settings import *
+import os
+from main import app
+
 class MapData(object):
 	def __init__(self, width, height):
 		self.width = width
@@ -9,7 +13,7 @@ class Layer(object):
 		self.layerIndex = layerIndex
 		
 		#Create gruop of tiles for this layer
-		self.tiles = pygame.sprite.Group()
+		self.tiles = pg.sprite.Group()
 		
 		#Create tiles in the right position for each layer
 		for x in range(self.mapObject.width):
@@ -22,12 +26,29 @@ class Layer(object):
 	def draw( self, screen ):
 		self.tiles.draw(screen)
 
+# Base class for anything that is within the world
+# Handles displaying the object on the screen based on a given viewpoint
+# Viewpoint can be any class with an offset and angle
+class WorldObject:
+	def __init__( self, pos ):
+		self.pos = vec2( pos )
+		self.screen_pos = vec2( 0 )
+		self.viewpoint = None
+
+	def update( self ):
+		self.calculate_viewpoint_position( self.viewpoint )
+
+	def calculate_viewpoint_position( self ):
+		view_pos = self.pos - self.viewpoint.offset
+		view_pos = view_pos.rotate_rad( self.viewpoint.angle )
+		self.screen_pos = view_pos + CENTER
+
 #Tile class with an image, x and y
-class Tile(pygame.sprite.Sprite):
-	def __init__( self, material, x, y ):
-		pygame.sprite.Sprite.__init__(self)
-		
-		self.image = material.image
+class Tile( pg.sprite.Sprite, WorldObject ):
+	def __init__( self, material, pos ):
+		super().__init__( pos )
+
+		self.sprite.image = material.image
 		self.rect = self.image.get_rect()
 		self.rect.x = x
 		self.rect.y = y
@@ -35,6 +56,6 @@ class Tile(pygame.sprite.Sprite):
 class Material:
 	def __init__( self, file ):
 		if( os.path.isfile(file) ):
-			self.image = pygame.image.load( file )
+			self.image = pg.image.load( file )
 		else:
-			self.image = pygame.image.load( "assets/materials/missing.png" )
+			self.image = pg.image.load( "assets/materials/missing.png" )
